refactor(java17): Backend initial refactor
Start to remove the redundancy on naming. First both src and svc
directories have their src- and svc- prefixes respectively removed.
Common now is part of the same base, instead of being isolated.

All dependencies and conflicts of build are solved.

At this moment the functionality still not checked, but no code
itself was modified.

Signed-off-by: Helio Chissini de Castro <heliocastro@gmail.com>